Korrektur im compresslog.sh

+++ docker/2.2.11/sources/compresslog.sh
@@ -1,9 +1,11 @@
 #!/bin/sh
-#echo "compresslog.sh"
-#echo "old log: ${1} new log ${2}"
 
-if [ "${1}" != "${2}" ]; then
-	echo "compressing logfile ${2}"
-	tar -czf "${2}.tar.gz" "${2}"
-	rm "${2}"
+if [ -n "$2" ]; then
+    if [ "${1}" != "${2}" ]; then
+	    echo "compressing logfile ${2}"
+    	tar -czf "${2}.tar.gz" "${2}"
+	    rm "${2}"
+    fi
+else
+    echo "no old logfile to compress"
 fi
